Backend-HiSparkAgent/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
requests
websockets
openai>=1.3.0
paho-mqtt>=1.6.1